changed structure w/ react router to allow for multipage, started work on integral
