+++ package.json
@@ -8,11 +8,14 @@
     "@testing-library/jest-dom": "^5.12.0",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
+    "@types/d3": "^7.0.0",
     "@types/jest": "^26.0.23",
     "@types/node": "^12.20.13",
     "@types/react": "^17.0.6",
     "@types/react-dom": "^17.0.5",
+    "@types/react-router-dom": "^5.3.0",
     "create-react-app": "^4.0.3",
+    "d3": "^7.1.1",
     "function-plot": "^1.22.7",
     "katex": "^0.13.11",
     "math.js": "^1.1.46",
@@ -22,6 +25,7 @@
     "react-dom": "^17.0.2",
     "react-mathjax": "^1.0.1",
     "react-mathjax2": "0.0.2",
+    "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3",
     "typescript": "^4.2.4",
     "web-vitals": "^1.1.2"
